Successful automated printing screenshot using shortcut
Shortcut cmdorctrl+shift+plus activates a shortcut listener in main, in the call back it sends a message to renderer a async message or request to take picture and send the dataURL as a symbol of the pixels captured from the canvas, this dataurl is caught by the async listener in main which creates a native image from dataurl, native image converts to png and outputs a buffer, a fs wrtiefilesync writes the file using path module, directory keyword and Date info. Several unsuccessful writes are included as pngs. many changes in api calls were made. downloadURL accepts Dataurls but asks for a prompt. editing the session on - will download method only complicates timing of actions and applies to a download item which is a object whose state can be checked to reflect status of remote download. fs write can not be done with combination of request since it  wont take dataurls.... odd since the default version works fine.